Purge default callback
