+++ lua/smoothcursor/callbacks/callback.lua
@@ -1,0 +1,1 @@
+-- Define callback global variable
